vhost: remove vhost_session_send_event()
Backends will now use spdk_thread_send_msg(), which is
much simpler. We'll now expect the vsession object not
to be removed before our event is processed, but that's
ensured by design. After all, we're waiting on a semaphore
on the DPDK thread, so DPDK can't possibly execute the
callback for terminating a connection.

Change-Id: Id26d268482b22e69f5ed7f7f5b16e860515332a9
Signed-off-by: Darek Stojaczyk <dariusz.stojaczyk@intel.com>