vhost: fail to stop session if it's not started
This can't technically happen unless DPDK is broken
and calls the callback for no reason, but we should
not break if that happens. Currently, completing the
stop without any error will decrease the started
sessions refcount and possibly make it smaller than 0.

Do the same for starting a session if it's already
started.

Change-Id: I8ab671ee3619e0aba4d621d9b7a768f8957890ef
Signed-off-by: Darek Stojaczyk <dariusz.stojaczyk@intel.com>