vhost: introduce spdk-specific vhost socket ops
They're basically a wrapper for rte_vhost's ops that
do the same. The difference is that our callbacks
are asynchronous (and are also called on spdk vhost's
init thread). Eventually we would like to put all
the code that needs to operate on the DPDK thread inside
rte_vhost_compat.c, so that there's no confusion which
data structures we can access where.

Change-Id: I957e51c7712f5bf22175c139c79a98ba39e4075f
Signed-off-by: Darek Stojaczyk <dariusz.stojaczyk@intel.com>